requirements.txt
streamlit
pandas
numpy
scipy
requests
google-generativeai
elevenlabs
tensorflow-cpu
scikit-learn
joblib
Pillow